requirements.txt
asgiref==3.8.1
asttokens==3.0.0
decorator==5.2.1
Django==5.1.6
djangorestframework==3.15.2
djangorestframework-jwt==1.11.0
djangorestframework_simplejwt==5.5.0
executing==2.2.0
ipdb==0.13.13
ipython==8.32.0
jedi==0.19.2
matplotlib-inline==0.1.7
parso==0.8.4
pexpect==4.9.0
prompt_toolkit==3.0.50
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
PyJWT==1.7.1
sqlparse==0.5.3
stack-data==0.6.3
traitlets==5.14.3
wcwidth==0.2.13
